refactor(synchronizer): rework files structure and types
